+++ uni_modules/vk-unicloud/components/vk-data-verification-code/vk-data-verification-code.vue
@@ -7,6 +7,51 @@
 </template>
 
 <script>
+
+const localeObj = {
+	"zh-Hans":{
+		"startText":"获取验证码",
+		"changeText":"X秒重新获取",
+		"endText":"重新获取",
+		"tryAgainInSeconds":"秒后再重试",
+		"pleaseEnterTheCorrectMobileNumber":"请输入正确的手机号码",
+		"sending":"发送中...",
+		"verificationCodeSent":"验证码已发送",
+		"triggerDayLevelFlowControl":"触发天级流控",
+		"pleaseTryAgainTomorrow":"短信发送频繁，请明日再试！",
+		"pleaseTryAgainIn1Hour":"短信发送频繁，请过1小时后再试！",
+		"triggerMinuteLevelFlowControl":"触发分钟级流控",
+		"pleaseTryAgainLater":"短信发送频繁，请稍后再试！",
+	},
+	"zh-Hant":{
+		"startText":"獲取驗證碼",
+		"changeText":"X秒重新獲取",
+		"endText":"重新獲取",
+		"tryAgainInSeconds":"秒後再重試",
+		"pleaseEnterTheCorrectMobileNumber":"請輸入正確的手機號碼",
+		"sending":"發送中...",
+		"verificationCodeSent":"驗證碼已發送",
+		"triggerDayLevelFlowControl":"觸發天級流控",
+		"pleaseTryAgainTomorrow":"短信發送頻繁，請明日再試！",
+		"pleaseTryAgainIn1Hour":"短信發送頻繁，請過1小時後再試！",
+		"triggerMinuteLevelFlowControl":"觸發分鐘級流控",
+		"pleaseTryAgainLater":"短信發送頻繁，請稍後再試！",
+	},
+	"en":{
+		"startText":"Get code",
+		"changeText":"X second reacquire",
+		"endText":"Reacquire",
+		"tryAgainInSeconds":"Try again in seconds",
+		"pleaseEnterTheCorrectMobileNumber":"Please enter the correct mobile number",
+		"sending":"Sending...",
+		"verificationCodeSent":"Verification code sent",
+		"triggerDayLevelFlowControl":"Trigger day level flow control",
+		"pleaseTryAgainTomorrow":"SMS sent frequently, please try again tomorrow!",
+		"pleaseTryAgainIn1Hour":"SMS sent frequently, please try again in 1 hour",
+		"triggerMinuteLevelFlowControl":"Trigger minute level flow control",
+		"pleaseTryAgainLater":"SMS sent frequently, please try again later",
+	}
+};
 /**
  * verificationCode 验证码输入框
  * @property {String} mode 模式，mobile 手机验证码 custom 自定义验证码 默认 mobile
@@ -53,17 +98,17 @@ export default {
 		// 尚未开始时提示
 		startText: {
 			type: String,
-			default: '获取验证码'
+			default: ''
 		},
 		// 正在倒计时中的提示
 		changeText: {
 			type: String,
-			default: 'X秒重新获取'
+			default: ''
 		},
 		// 倒计时结束时的提示
 		endText: {
 			type: String,
-			default: '重新获取'
+			default: ''
 		},
 		// 是否在H5刷新或各端返回再进入时继续倒计时
 		keepRunning: {
@@ -81,8 +126,17 @@ export default {
 			secNum: this.seconds,
 			timer: null,
 			canGetCode: true, // 是否可以执行验证码操作
-			tips: ''
+			tips: '',
+			locale:{}, // 语言
 		};
+	},
+	created(){
+		if (typeof vk !== "undefined") {
+			let locale = vk.pubfn.getLocale();
+			this.locale = localeObj[locale];
+		} else {
+			this.locale = localeObj["zh-Hans"];
+		}
 	},
 	mounted() {
 		this.checkKeepRunning();
@@ -95,11 +149,23 @@ export default {
 			}
 		}
 	},
+	// 计算属性
+	computed:{
+		startTextCom(){
+			return this.startText || this.locale.startText || "获取验证码";
+		},
+		changeTextCom(){
+			return this.changeText || this.locale.changeText || "X秒重新获取";
+		},
+		endTextCom(){
+			return this.endText || this.locale.endText || "重新获取";
+		}
+	},
 	methods: {
 		checkKeepRunning() {
 			// 获取上一次退出页面(H5还包括刷新)时的时间戳，如果没有上次的保存，此值可能为空
 			let lastTimestamp = Number(uni.getStorageSync(this.uniqueKey + '_$uCountDownTimestamp'));
-			if (!lastTimestamp) return this.changeEvent(this.startText);
+			if (!lastTimestamp) return this.changeEvent(this.startTextCom);
 			// 当前秒的时间戳
 			let nowTimestamp = Math.floor(+new Date() / 1000);
 			// 判断当前的时间戳，是否小于上一次的本该按设定结束，却提前结束的时间戳
@@ -112,7 +178,7 @@ export default {
 				this.start();
 			} else {
 				// 如果不存在需要继续上一次的倒计时，执行正常的逻辑
-				this.changeEvent(this.startText);
+				this.changeEvent(this.startTextCom);
 			}
 		},
 		// 开始倒计时
@@ -125,16 +191,16 @@ export default {
 			this.$emit('start');
 			this.canGetCode = false;
 			// 这里放这句，是为了一开始时就提示，否则要等setInterval的1秒后才会有提示
-			this.changeEvent(this.changeText.replace(/x|X/, this.secNum));
+			this.changeEvent(this.changeTextCom.replace(/x|X/, this.secNum));
 			this.setTimeToStorage();
 			this.timer = setInterval(() => {
 				if (--this.secNum) {
 					// 用当前倒计时的秒数替换提示字符串中的"x"字母
-					this.changeEvent(this.changeText.replace(/x|X/, this.secNum));
+					this.changeEvent(this.changeTextCom.replace(/x|X/, this.secNum));
 				} else {
 					clearInterval(this.timer);
 					this.timer = null;
-					this.changeEvent(this.endText);
+					this.changeEvent(this.endTextCom);
 					this.secNum = this.seconds;
 					this.$emit('end');
 					this.canGetCode = true;
@@ -146,7 +212,7 @@ export default {
 			this.canGetCode = true;
 			clearInterval(this.timer);
 			this.secNum = this.seconds;
-			this.changeEvent(text || this.endText);
+			this.changeEvent(text || this.endTextCom);
 		},
 		changeEvent(text) {
 			this.codeChange(text);
@@ -177,7 +243,7 @@ export default {
 			let that = this;
 			let vk = uni.vk;
 			if (!that.canGetCode) {
-				vk.toast(`${that.secNum}秒后再重试`, 'none');
+				vk.toast(`${that.secNum}${that.locale.tryAgainInSeconds}`, 'none');
 				return;
 			}
 			if (that.mode === "custom") {
@@ -187,10 +253,10 @@ export default {
 			let mobile = that.mobile;
 			let type = that.type;
 			if (!vk.pubfn.test(mobile, 'mobile')) {
-				vk.toast('请输入正确的手机号码', 'none');
+				vk.toast(that.locale.pleaseEnterTheCorrectMobileNumber, 'none');
 				return;
 			}
-			that.tips = "发送中...";
+			that.tips = that.locale.sending;
 			vk.userCenter.sendSmsCode({
 				needAlert: false,
 				data: {
@@ -198,21 +264,21 @@ export default {
 					type
 				},
 				success: function(data) {
-					vk.toast('验证码已发送');
+					vk.toast(that.locale.verificationCodeSent);
 					that.start();
 				},
 				fail: function(err) {
-					that.tips = that.startText;
+					that.tips = that.startTextCom;
 					if (err.errMsg && err.errMsg.indexOf('触发天级流控') > -1) {
-						vk.alert('短信发送频繁，请明日再试！');
+						vk.alert(that.locale.pleaseTryAgainTomorrow);
 					} else if (err.errMsg && err.errMsg.indexOf('触发小时级流控') > -1) {
-						vk.alert('短信发送频繁，请过1小时后再试！');
+						vk.alert(that.locale.pleaseTryAgainIn1Hour);
 					} else if (err.errMsg && err.errMsg.indexOf('触发分钟级流控') > -1) {
-						vk.alert('短信发送频繁，请稍后再试！');
+						vk.alert(that.locale.pleaseTryAgainLater);
 					} else if (err.msg) {
 						vk.alert(err.msg);
 					} else {
-						vk.alert('短信发送频繁，请稍后再试！');
+						vk.alert(that.locale.pleaseTryAgainLater);
 					}
 				}
 			});